requirements.txt
absl-py==0.12.0
aiofiles==0.8.0
anyio==3.6.1
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
awscli==1.24.10
Babel==2.10.3
backcall==0.2.0
backports.zoneinfo==0.2.1
bleach==4.1.0
boto3==1.23.10
botocore==1.26.10
cached-property==1.5.2
cachetools==4.2.4
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
clang==5.0
colorama==0.4.4
contextvars==2.4
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.4
dnspython==2.2.1
docutils==0.16
entrypoints==0.4
eventlet==0.33.1
fiftyone-brain==0.8.2
fiftyone-db==0.3.0
fiftyone-db-rhel7==0.3.0
flatbuffers==1.12
future==0.18.2
gast==0.4.0
glob2==0.7
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.3
greenlet==1.1.2
grpcio==1.46.3
h11==0.12.0
h5py==3.1.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
imageio==2.15.0
immutables==0.18
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.17.2
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
json5==0.9.8
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-server==1.13.1
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.1.1
kaleido==0.2.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
mongoengine==0.20.0
motor==2.5.1
munkres==1.1.4
nbclassic==0.3.5
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
ndjson==0.3.1
nest-asyncio==1.5.5
networkx==2.5.1
notebook==6.4.10
numpy==1.19.5
oauthlib==3.2.0
open3d==0.9.0.0
opencv-python==4.5.5.64
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
patool==1.12
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==4.14.3
pprintpp==0.4.0
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
pylops==1.18.2
pymongo==3.12.3
pyparsing==3.0.9
pyproximal==0.2.0
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==23.2.0
ranger==0.10
requests==2.27.1
requests-oauthlib==1.3.1
retrying==1.3.3
rfc3986==1.5.0
rsa==4.7.2
s3transfer==0.5.2
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
Send2Trash==1.8.0
six==1.15.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
tabulate==0.8.10
termcolor==1.1.0
terminado==0.12.1
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2020.9.3
torch==1.10.1+cu113
torchaudio==0.10.1
torchsummary==1.5.1
torchvision==0.11.2
tornado==6.1
tqdm==4.64.0
traitlets==4.3.3
typing-extensions==3.7.4.3
tzdata==2022.1
tzlocal==4.2
universal-analytics-python3==1.1.1
urllib3==1.26.9
voxel51-eta==0.6.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.0.3
widgetsnbextension==3.6.1
wrapt==1.12.1
xmltodict==0.13.0
zipp==3.6.0
